Export applyValueToSides util function from BoxControl
This could avoid some code duplication for new components around spacing size presets.
