+++ packages/components/src/box-control/index.js
@@ -173,3 +173,5 @@ export default function BoxControl( {
 		</Root>
 	);
 }
+
+export { applyValueToSides } from './utils';
